Update uint256 to uint64 for depositNonce

+++ contracts/interfaces/IDepositExecute.sol
@@ -12,7 +12,7 @@ interface IDepositExecute {
         @param depositer Address of account making the deposit in the Bridge contract.
         @param data Consists of additional data needed for a specific deposit.
      */
-    function deposit(uint8 destinationChainID, uint256 depositNonce, address depositer, bytes calldata data) external;
+    function deposit(uint8 destinationChainID, uint64 depositNonce, address depositer, bytes calldata data) external;
 
     /**
         @notice It is intended that proposals are executed by the Bridge contract.
